P0/requirements.txt
scikit_image==0.17.2
numpy==1.20.0
matplotlib==3.3.4
Pillow==8.1.0
skimage==0.0
